Make a types module and fix type errors
