+++ tests/sentry/ingest/test_userreport.py
@@ -2,6 +2,7 @@ from unittest.mock import Mock
 
 from django.utils import timezone
 
+from sentry.feedback.lib.types import UserReportDict
 from sentry.feedback.usecases.create_feedback import (
     UNREAL_FEEDBACK_UNATTENDED_MESSAGE,
     FeedbackCreationSource,
@@ -73,7 +74,7 @@ def test_save_user_report_returns_instance(default_project, monkeypatch):
     )
 
     # Test data
-    report = {
+    report: UserReportDict = {
         "event_id": "123456",
         "name": "Test User",
         "email": "test@example.com",
@@ -91,7 +92,7 @@ def test_save_user_report_returns_instance(default_project, monkeypatch):
 @django_db_all
 def test_save_user_report_denylist(default_project, monkeypatch):
     monkeypatch.setattr("sentry.ingest.userreport.is_in_feedback_denylist", lambda org: True)
-    report = {
+    report: UserReportDict = {
         "event_id": "123456",
         "name": "Test User",
         "email": "test@example.com",
@@ -116,7 +117,7 @@ def test_save_user_report_filters_large_message(default_project, monkeypatch):
     if not max_length:
         assert False, "Missing max_length for UserReport comments field!"
 
-    report = {
+    report: UserReportDict = {
         "event_id": "123456",
         "name": "Test User",
         "email": "test@example.com",
@@ -145,7 +146,7 @@ def test_save_user_report_shims_if_event_found(default_project, monkeypatch):
     mock_shim_to_feedback = Mock()
     monkeypatch.setattr("sentry.ingest.userreport.shim_to_feedback", mock_shim_to_feedback)
 
-    report = {
+    report: UserReportDict = {
         "event_id": event.event_id,
         "name": "Test User",
         "email": "test@example.com",
@@ -176,7 +177,7 @@ def test_save_user_report_does_not_shim_if_event_found_but_source_is_new_feedbac
     mock_shim_to_feedback = Mock()
     monkeypatch.setattr("sentry.ingest.userreport.shim_to_feedback", mock_shim_to_feedback)
 
-    report = {
+    report: UserReportDict = {
         "event_id": event.event_id,
         "name": "Test User",
         "email": "test@example.com",
